Move PropertySyntaxVisitor to its own file
